+++ udata_front/theme/gouvfr/assets/js/pages/OrganizationPublishingForm/OrganizationPublishingForm.stories.ts
@@ -1,4 +1,4 @@
-import { rest } from 'msw';
+import { delay, http, HttpResponse } from 'msw';
 import OrganizationPublishingForm from './OrganizationPublishingForm.vue';
 import type { Meta, StoryObj } from "@storybook/vue3";
 import { Dataset, NewDataset, Organization } from '@etalab/data.gouv.fr-components';
@@ -15,13 +15,16 @@ const args = {};
 export const Form: StoryObj<typeof meta> = {
   parameters: {
     msw: [
-      rest.post('*/api/1/organizations/', async (req, res, ctx) => {
-        const body: Organization = await req.json();
+      http.post('*/api/1/organizations/', async ({ request }) => {
+        const body: Organization = await request.json();
         const organization: Organization = {...body, id: "someId", last_update: new Date()};
-        return res(ctx.delay(), ctx.json(organization));
+        await delay();
+        return HttpResponse.json(organization);
       }),
-      rest.post('*/api/1/organizations/:organizationId/logo', async (_req, res, ctx) => {
-        return res(ctx.delay(), ctx.status(200), ctx.json({"image": "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png", "success": true}));
+      http.post('*/api/1/organizations/:organizationId/logo', async () => {
+        return HttpResponse.json({"image": "https://demo-static.data.gouv.fr/avatars/84/3194d831264f769fa817e58813d413-100.png", "success": true}), {
+          status: 200,
+        };
       }),
     ],
   },
@@ -40,13 +43,17 @@ export const Form: StoryObj<typeof meta> = {
 export const FormWithFailedRequests: StoryObj<typeof meta> = {
   parameters: {
     msw: [
-      rest.post('*/api/1/organizations/', async (req, res, ctx) => {
-        const body: NewDataset = await req.json();
+      http.post('*/api/1/organizations/', async ({ request }) => {
+        const body: NewDataset = await request.json();
         const organization: Dataset = {...body, id: "someId", last_update: new Date()};
-        return res(ctx.delay(), ctx.json(organization));
+        await delay();
+        return HttpResponse.json(organization);
       }),
-      rest.post('*/api/1/organizations/:organizationId/upload', async (_req, res, ctx) => {
-        return res(ctx.delay(), ctx.status(400), ctx.json({error: "Chunk size mismatch"}));
+      http.post('*/api/1/organizations/:organizationId/upload', async () => {
+        await delay();
+        return HttpResponse.json({error: "Chunk size mismatch"}, {
+          status: 400,
+        });
       }),
     ],
   },