+++ udata_front/theme/gouvfr/datagouv-components/src/components/Owner/index.ts
@@ -1,0 +1,4 @@
+import OwnerType from "./OwnerType.vue";
+import OwnerTypeIcon from "./OwnerTypeIcon.vue";
+
+export { OwnerType, OwnerTypeIcon };