+++ udata_front/theme/gouvfr/datagouv-components/src/helpers/throwOnNever.ts
@@ -1,0 +1,3 @@
+export function throwOnNever(_: never, message: string): never {
+  throw new Error(message);
+}