+++ udata_front/theme/gouvfr/datagouv-components/vite.config.ts
@@ -48,5 +48,8 @@ export default defineConfig(config => {
       },
       sourcemap: true,
     },
+    optimizeDeps: {
+      exclude: ["oh-vue-icons/icons"]
+    }
   };
 });