+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add organization edit to new admin [#412](https://github.com/datagouv/udata-front/pull/412)
 - Fix markdown tables and use DSFR ones [#432](https://github.com/datagouv/udata-front/pull/432)
+- Add organization selector to dataset publishing [#438](https://github.com/datagouv/udata-front/pull/438)
 - Add reports [#436](https://github.com/datagouv/udata-front/pull/436)
 
 ## 5.0.1 (2024-06-12)